Implement findByPropertyIsTrue findByPropertyIsFalse
