+++ spring-data-neo4j/src/test/java/org/springframework/data/neo4j/examples/restaurants/repo/RestaurantRepository.java
@@ -71,7 +71,7 @@ public interface RestaurantRepository extends GraphRepository<Restaurant> {
 
 	List<Restaurant> findByHalalIsTrue();
 
-	List<Restaurant> findByHalalIsNotTrue();
+	List<Restaurant> findByHalalIsFalse();
 
 }
